A gross and dirty hack to fix a bug in JaCoCo, see code snippet below:
```
public static void branchWithException(int input) {
        if (input == 0) {
            System.out.println("Input is 0");
        } else {
            alwaysThrowsException();
        }
    }

    public static void alwaysThrowsException() {
        throw new IllegalStateException();
    }
```

In this case, JaCoCo will never report that both sides of the branch in `branchWithException` are covered, since it won't put a probe before the call to `alwaysThrowsException` (only after it!). There is already some code in JaCoCo to collect coverage of instructions that are predecessors to method invocations, but that code won't insert a probe between a jump target (the else branch) and the method invocation. There's surely a much better way to patch this, but this is a quick and dirty fix that will add a no-op instruction after branch targets, allowing JaCoCo's existing code to determine to insert a probe there.